+++ src/components/routing/routers/GameRouter.js
@@ -1,5 +1,4 @@
 import {Redirect, Route} from "react-router-dom";
-import Game from "components/views/Game";
 import PropTypes from 'prop-types';
 import Dashboard from "../../views/Dashboard";
 
@@ -10,7 +9,7 @@ const GameRouter = props => {
     return (
         <div style={{display: 'flex', flexDirection: 'column'}}>
             <Route exact path={`${props.base}/dashboard`}>
-                <Game/>
+                <Dashboard/>
             </Route>
             <Route exact path={`${props.base}`}>
                 <Redirect to={`${props.base}/dashboard`}/>
