+++ src/components/views/Dashboard.js
@@ -1,9 +1,9 @@
 import "styles/views/Dashboard.scss";
 import "styles/_theme.scss";
-import SideBar from "../ui/SideBar";
+import SideBar from "../ui/Dashboard_ui/SideBar";
 import * as React from "react";
 
-import TableUserOverview from "../ui/TableUserOverview";
+import TableUserOverview from "../ui/Dashboard_ui/TableUserOverview";
 
 const Dashboard = () => {
