+++ src/components/views/Login.js
@@ -7,7 +7,6 @@ import 'styles/views/Login.scss';
 import BaseContainer from "components/ui/BaseContainer";
 import PropTypes from "prop-types";
 import TextField from '@mui/material/TextField';
-import Header from "./Header";
 import AppRouter from "../routing/routers/AppRouter";
 
 const FormField = props => {
