+++ src/styles/ui/Dashboard_ui/SideBar.scss
@@ -1,4 +1,4 @@
-@import 'src/styles/_theme.scss';
+@import 'src/styles/theme';
 
 
 .sidebar {
