+++ src/styles/ui/Dashboard_ui/TableUserOverview.scss
@@ -1,4 +1,4 @@
-@import 'src/styles/_theme.scss';
+@import 'src/styles/theme';
 
 $paddingField: 9px 9px 9px 11px;
