--- src/styles/ui/ReactLogo.scss
@@ -1,12 +1,0 @@
-@import 'src/styles/_theme.scss';
-
-.react-logo {
-  fill: $background;
-  animation: spin 20s linear infinite;
-}
-
-@keyframes spin {
-  100% {
-    transform: rotate(360deg);
-  }
-}
