--- src/styles/views/Header.scss
@@ -1,13 +1,0 @@
-@import 'src/styles/_theme.scss';
-
-.header {
-  &.container {
-    @include center;
-  }
-
-  &.title {
-    font-weight: bold;
-    color: $textColor;
-    text-align: center;
-  }
-}
